Add Godep and vendor solution to this project
Godep is a package / dependency manager that is used to help manage
external library dependencies. Look at Godeps/Readme for more info.

Future checkins will need to run `godep save ./...` to update the
list of depencies before checking in. The CI system will pickup any
failures until we automate this as part of the checkin process.